more cleanup
